[Fix] 뽑기 탭으로 복귀 시, 애니메이션 계속 실행되는 오류 수정
