+++ composeApp/src/commonMain/kotlin/com/kus/kustaurant/navigation/MainScreen.kt
@@ -196,7 +196,6 @@ fun MainScreen(
             drawNavGraph(
                 onSearchClick = { rootNavController.navigate(Search) },
                 onAlarmClick = {},
-                onShowMessage = onShowMessage,
                 navigateToDrawResult = { route -> mainNavController.navigate(route) },
                 onBackClick = { mainNavController.popBackStack() },
             )
