+++ shared/core/designSystem/src/commonMain/composeResources/drawable/img_no_result.png
