+++ shared/feature/draw/src/commonMain/kotlin/com/kus/feature/draw/ui/result/DrawResultUiState.kt
@@ -1,5 +1,6 @@
 package com.kus.feature.draw.ui.result
 
+import UiState
 import com.kus.domain.draw.model.DrawRestaurant
 import com.kus.shared.domain.model.tier.filter.Cuisine
 import com.kus.shared.domain.model.tier.filter.Location
@@ -7,10 +8,9 @@ import com.kus.shared.domain.model.tier.filter.Location
 data class DrawResultUiState(
     val selectedLocations: Set<Location> = emptySet(),
     val selectedCuisines: Set<Cuisine> = emptySet(),
-    val restaurants: List<DrawRestaurant> = emptyList(),
-    val isLoading: Boolean = false,
-    val toastMessage: String? = null,
+    val restaurantState: UiState<List<DrawRestaurant>> = UiState.Idle,
+    val randomIndex: Int? = null,
+    val sameDrawCnt: Int = 0,
     val initialized: Boolean = false,
-    val randomIndex : Int? = null,
-    val sameDrawCnt : Int = 0,
+    val hasPlayedDrawAnimation: Boolean = false,
 )
